Get from DB working for messages
