+++ app/controllers/users.server.controller.js
@@ -11,4 +11,10 @@ var _ = require('lodash');
 module.exports = _.extend(
 	require('./users/users.authentication.server.controller'),
 	require('./users/users.authorization.server.controller')
+
+
+
 );
+
+
+
