+++ app/views/layout.server.view.html
@@ -65,10 +65,12 @@
 	<!--Livereload script rendered -->
 	<script type="text/javascript" src="http://{{request.hostname}}:35729/livereload.js"></script>
 	{% endif %}
-</body>
-		
 		<script type='text/javascript' src='/socket.io/socket.io.js'></script>
-		<script type="text/javascript" src ="/socket.io/client.js"></script>
+		<script>
+			var socket = io();
+		</script>
+
+</body>
 		<script src="http://ajax.googleapis.com/ajax/libs/jquery/1.11.3/jquery.min.js"></script>
 
 </html>
